Treat an empty CI-identifying env variable as non-existing
When running Nuke from inside a LINUX docker container, we tend to pass the TEAMCITY_VERSION from the host to the container. However, that will set the value to an empty string, even though the variable did not exist on the host. This is fixed for Team City and all the other CI providers.