Resolves an issue where npm task would log warnings as errors
NPM outputs warnings to the error stream, some build systems fail because they assume there was build errors.

A custom logger was implemented in #419 but it checks for "npmWARN" instead of the current "npm WARN" (note the space). I am not sure if that was originally a typo or if it changed at some NPM version. This PR should handle both cases.