Fix #1181 UnityTasks.Unity throws exception in AbsolutePathExtensions.Exists on macOS
I considered adding tests for this issue, but found that it would require sweeping changes to decouple the platform and system logic. A more simplified test would be one that would only detect the issue when actually running on macOS and not on Windows, but I don't yet know if the Nuke project has anything set up for platform testing to make such tests feasible.

For now, the change is simple enough to warrant that we go without a test, I believe.