Also call DetectUnityVersion when in UnityTasks.UnityRunTests
The generic PreProcess method on line 111 also handles Unity runs that don't directly relate to a project (e.g. license activation). However, for the tasks which deal with projects, we want them to auto-detect the Unity editor version from the project. Previously this was only implemented when starting the general UnityTasks.Unity method, but now it's also used when running tests.